requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
dj-database-url>=3.0,<4.0
python-dotenv>=1.1,<2.0
psycopg[binary]>=3.2,<4.0
gunicorn>=21.2,<24.0
# 2FA Dependencies
pyotp>=2.9,<3.0
qrcode[pil]>=7.4,<8.0
pillow>=11.0,<12.0

# Rate Limiting
django-ratelimit>=4.1,<5.0

# CORS Headers
django-cors-headers>=4.9,<5.0

# API Documentation
drf-spectacular>=0.27.0,<0.30.0

# Testing
pytest>=8.0,<10.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<8.0
factory-boy>=3.3,<4.0
faker>=22.0,<41.0
